Actualizar

+++ front/src/app/circuit-editor/circuit-editor.component.ts
@@ -8,6 +8,9 @@ import { QiskitService } from '../qiskit.service';
 import { QubitsConfigurationService } from '../qubits-configuration.service';
 import { EdCircuitsService } from '../ed-circuits.service';
 
+import { Backend } from '../deterministic/Backend';
+import { TranspileService } from '../transpile.service';
+
 @Component({
   selector: 'app-circuit-editor',
   templateUrl: './circuit-editor.component.html',
@@ -37,11 +40,25 @@ export class CircuitEditorComponent {
   customizedGates : EdGate[] = [];
 
   mensajeTemporal: string = '';
+  mensajeTemporal2: string = '';
   searchQuery: string = "";
   modalCodigo: boolean = false;
   mostrarModalCrear: boolean = false;
+  gateMselected: boolean = false;
+  gatesModal: boolean = false;
+  showDeleteModal: boolean = false;
+  deleteIndex: number = 0;
+  gateToDelete?: EdGate;
+  modalCodigoGate: boolean = false;
+  modalTranspile: boolean = false;
 
-  constructor(public manager : ManagerService, private qiskitService : QiskitService, private qubitsConfigurationService: QubitsConfigurationService, private circuitsService : EdCircuitsService) {
+  circuitName: string = '';
+  transpiledCode: string = '';
+  availableBackends: Backend[] = [];
+  selectedBackends: Backend[] = [];
+
+  constructor(public manager : ManagerService, private qiskitService : QiskitService, private qubitsConfigurationService: QubitsConfigurationService, 
+    private circuitsService : EdCircuitsService, public transpileService: TranspileService) {
     this.qiskitService.getCustomizedGates().subscribe(
       gates => {
         for (let i=0; i<gates.length; i++) {
@@ -67,6 +84,15 @@ export class CircuitEditorComponent {
   }
 
   ngOnInit() {
+
+    this.transpileService.getBackends().subscribe(backends => {
+      this.availableBackends = backends;
+    });
+    
+    
+    this.selectedBackends = JSON.parse(localStorage.getItem('selectedBackends') || '[]');
+    this.availableBackends = JSON.parse(localStorage.getItem('availableBackends') || '[]');
+
     this.manager.selectedTemplate = this.manager.templates[0];
     this.qubitsConfigurationService.getQubitConfigurationNames().subscribe(
         qubitsConfiguration => {
@@ -263,6 +289,7 @@ export class CircuitEditorComponent {
   }
 
   selectGate(gate: EdGate) {
+    this.gateMselected = false;
     if (this.selectedGate === gate) {
       this.selectedGate = undefined; // Deseleccionar si ya estaba seleccionada
     } else {
@@ -273,14 +300,23 @@ export class CircuitEditorComponent {
   selectMeasurementGate() {
     if (!this.circuit)
       return;
-    this.selectedGate = new EdGate('M', 1);
-    this.selectedGate.description = 'Measure this qubit';
-    this.selectedGate.code = 'circuit.measure(' + this.circuit.qubits.length + ', ' + this.circuit.qubits.length + ')';
+    if (this.gateMselected) {
+      this.gateMselected = false;
+      this.selectedGate = undefined;
+    } else {
+      this.selectedGate = new EdGate('M', 1);
+      this.gateMselected = true;
+      this.selectedGate.description = 'Measure this qubit';
+      this.selectedGate.code = 'circuit.measure(' + this.circuit.qubits.length + ', ' + this.circuit.qubits.length + ')';
+    }
+    
   }
 
   createCustomizedGate() {
     this.selectedGate = new EdGate('', 1);
+    console.log("Selected gate: ", this.selectedGate);
     this.creatingNewGate = true;
+    //this.selectedGate = undefined;
 
     this.mostrarModalCrear = true;
   }
@@ -289,7 +325,13 @@ export class CircuitEditorComponent {
     this.creatingNewGate = false;
     this.qiskitService.saveGate(this.selectedGate!).subscribe(
         ok=> {
-          this.customizedGates.push(this.selectedGate!)
+          this.customizedGates.push(this.selectedGate!);
+          this.mensajeTemporal = 'Gate created successfully';
+          this.selectedGate = undefined;
+          this.mostrarModalCrear = false;
+          setTimeout(() => {
+            this.mensajeTemporal = '';
+          }, 2000);
         },
         error => {
           this.error = error.error.message
@@ -298,14 +340,22 @@ export class CircuitEditorComponent {
   }
 
   deleteFromServer() {
-    let option = confirm("Are you sure you want to delete this gate?");
+    /*let option = confirm("Are you sure you want to delete this gate?");
     if (!option || !this.selectedGate)
-      return;
+      return;*/
+    if (!this.selectedGate)
+      return
     this.creatingNewGate = false;
     this.qiskitService.deleteFromServer(this.selectedGate!).subscribe(
         ok=> {
           this.customizedGates = this.customizedGates.filter(g => g.name !== this.selectedGate!.name);
           this.selectedGate = undefined;
+          this.gateToDelete = undefined;
+          this.showDeleteModal = false;
+          this.mensajeTemporal = 'Gate deleted successfully';
+          setTimeout(() => {
+            this.mensajeTemporal = '';
+          }, 2000);
         },
         error => {
           this.error = error.error.message
@@ -429,7 +479,22 @@ export class CircuitEditorComponent {
     const codigo = this.code?.toString() || '';
     navigator.clipboard.writeText(codigo).then(() => {
       console.log('Código copiado al portapapeles');
-      alert('Code copied to clipboard');
+      this.mensajeTemporal2 = 'Code copied';
+      setTimeout(() => {
+          this.mensajeTemporal2 = '';
+      }, 1000);
+    }).catch(err => {
+      console.error('Error al copiar el código:', err);
+    });
+  }
+
+  copiarCodigoGate() {
+    const codigo = this.selectedGate?.code.toString() || '';
+    navigator.clipboard.writeText(codigo).then(() => {
+      this.mensajeTemporal2 = 'Code copied';
+      setTimeout(() => {
+          this.mensajeTemporal2 = '';
+      }, 1000);
     }).catch(err => {
       console.error('Error al copiar el código:', err);
     });
@@ -465,5 +530,68 @@ export class CircuitEditorComponent {
     }
     return false;
   }
+
+  openDeleteModal(gate: any, index: number) {
+    this.gateToDelete = gate;
+    this.deleteIndex = index;
+    this.showDeleteModal = true;
+  }
+
+  confirmDelete() {
+    this.deleteFromServer();
+    this.selectedGate = this.gateToDelete;
+  }
+
+  cancelDelete() {
+    this.gateToDelete = undefined;
+    this.showDeleteModal = false;
+  }
+
+  codeShowGates(gate: any) {
+    this.selectedGate = gate;
+    this.modalCodigoGate = true;
+  }
+
+  transpileCodigo() {
+    this.modalTranspile = true;
+  }
+
+  selectBackend(backend: Backend) {
+        this.selectedBackends.push(backend);
+        this.availableBackends = this.availableBackends.filter(b => b.name !== backend.name);
+        localStorage.setItem('selectedBackends', JSON.stringify(this.selectedBackends));
+        localStorage.setItem('availableBackends', JSON.stringify(this.availableBackends));
+      }
+    
+      deselectBackend(backend: Backend) {
+        this.availableBackends.push(backend);
+        this.selectedBackends = this.selectedBackends.filter(b => b.name !== backend.name);
+        localStorage.setItem('selectedBackends', JSON.stringify(this.selectedBackends));
+        localStorage.setItem('availableBackends', JSON.stringify(this.availableBackends));
+      }
+    
+      transpile() {
+        try{
+          const backendsToTranspile = this.selectedBackends.map(b => b.name);
+          if (this.code) {
+            this.transpileService.transpile(this.code, backendsToTranspile, this.circuitName).subscribe(result => {
+              this.transpiledCode = result;
+            });
+          }
+          
+          this.mensajeTemporal = 'The code will be transpiled.';
+          setTimeout(() => {
+            this.mensajeTemporal = '';
+          }
+          , 2000);
+        } catch (error) {
+          console.error('Error during transpilation:', error);
+          this.mensajeTemporal = 'Error during transpilation. Please try again.';
+          setTimeout(() => {
+            this.mensajeTemporal = '';
+          }, 2000);
+        }
+        
+      }
 
 }
